server/requirements.txt
astropy==5.1
bio==1.3.9
biopython==1.79
biothings-client==0.2.6
certifi==2022.6.15
cftime==1.6.1
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
decode==0.7.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
fonttools==4.37.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
morecopy==0.2.4
mygene==3.2.2
netCDF4==1.6.0
numpy==1.22.4
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyerfa==2.0.0.1
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.7.0
six==1.16.0
SQLAlchemy==1.4.40
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.0
typing-extensions==3.10.0.2
urllib3==1.26.12
Werkzeug==2.2.2
xarray==0.20.2
xarray-dataclasses==1.2.0
zipp==3.8.1
